requirements.txt
Django==4.2.21
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==10.4.0
django-cors-headers
